requirements.txt
flask
python-dotenv
newsapi-python
newspaper3k
